example of customHook
