+++ src/components/BackwardCounter.js
@@ -1,17 +1,8 @@
-import { useState, useEffect } from 'react';
-
 import Card from './Card';
+import useCounter from './hooks/use-count';
 
 const BackwardCounter = () => {
-  const [counter, setCounter] = useState(0);
-
-  useEffect(() => {
-    const interval = setInterval(() => {
-      setCounter((prevCounter) => prevCounter - 1);
-    }, 1000);
-
-    return () => clearInterval(interval);
-  }, []);
+  const counter = useCounter(false);
 
   return <Card>{counter}</Card>;
 };
